+++ .github/workflows/integration.yaml
@@ -1,4 +1,4 @@
-name: End to End Tests
+name: Integration Tests
 
 on:
   push:
@@ -32,7 +32,7 @@ jobs:
           name: k8s.snap
           path: k8s.snap
 
-  test-e2e:
+  test-integration:
     name: Test ${{ matrix.os }}
     strategy:
       matrix:
@@ -65,7 +65,7 @@ jobs:
           export TEST_SNAP="$PWD/build/k8s.snap"
           export TEST_SUBSTRATE=lxd
           export TEST_LXD_IMAGE=${{ matrix.os }}
-          cd tests/e2e && sg lxd -c 'tox -e e2e'
+          cd tests/integration && sg lxd -c 'tox -e integration'
 
   security-scan:
     name: Security scan
