+++ .github/workflows/python.yaml
@@ -22,4 +22,4 @@ jobs:
         run: pip install tox
       - name: Lint
         run: |
-          cd tests/e2e && tox -e lint
+          cd tests/integration && tox -e lint
