--- tests/e2e/tests/e2e_util/harness/__init__.py
@@ -1,18 +1,0 @@
-#
-# Copyright 2024 Canonical, Ltd.
-#
-from e2e_util.harness.base import Harness, HarnessError, Instance
-from e2e_util.harness.juju import JujuHarness
-from e2e_util.harness.local import LocalHarness
-from e2e_util.harness.lxd import LXDHarness
-from e2e_util.harness.multipass import MultipassHarness
-
-__all__ = [
-    HarnessError,
-    Harness,
-    Instance,
-    JujuHarness,
-    LocalHarness,
-    LXDHarness,
-    MultipassHarness,
-]
