+++ tests/integration/.copyright.tmpl
