+++ tests/integration/README.md
@@ -26,7 +26,7 @@ mv k8s_*.snap k8s.snap
 
 In general, all end to end tests will require specifying the local path to the snap package under test, using the `TEST_SNAP` environment variable. Make sure to specify the full path to the file.
 
-End to end tests are typically run with: `cd tests/e2e && tox -e e2e`
+End to end tests are typically run with: `cd tests/integration && tox -e integration`
 
 ### Running end to end tests on the local machine
 
@@ -34,7 +34,7 @@ End to end tests are typically run with: `cd tests/e2e && tox -e e2e`
 export TEST_SNAP=$PWD/k8s.snap
 export TEST_SUBSTRATE=local
 
-cd tests/e2e && tox -e e2e
+cd tests/integration && tox -e integration
 ```
 
 > *NOTE*: When running locally, end to end tests that create more than one instance will fail.
@@ -54,10 +54,10 @@ export TEST_SNAP=$PWD/k8s.snap
 export TEST_SUBSTRATE=lxd
 
 export TEST_LXD_IMAGE=ubuntu:22.04          # (optionally) specify which image to use for LXD containers
-export TEST_LXD_PROFILE=k8s-e2e             # (optionally) specify profile name to configure
+export TEST_LXD_PROFILE=k8s-integration     # (optionally) specify profile name to configure
 export TEST_SKIP_CLEANUP=1                  # (optionally) do not destroy machines after tests finish
 
-cd tests/e2e && tox -e e2e
+cd tests/integration && tox -e integration
 ```
 
 ### Running end to end tests on multipass VMs
@@ -79,7 +79,7 @@ export TEST_MULTIPASS_CPUS=4                # (optionally) specify how many cpus
 export TEST_MULTIPASS_MEMORY=2G             # (optionally) specify how much RAM each VM should have
 export TEST_MULTIPASS_DISK=10G              # (optionally) specify how much disk each VM should have
 
-cd tests/e2e && tox -e e2e
+cd tests/integration && tox -e integration
 ```
 
 Multipass can also be used to run the tests on Ubuntu Core:
@@ -89,18 +89,18 @@ export TEST_SNAP=$PWD/k8s.snap
 export TEST_SUBSTRATE=multipass
 export TEST_MULTIPASS_IMAGE=core20
 
-cd tests/e2e && tox -e e2e
+cd tests/integration && tox -e integration
 ```
 
 ### Running end to end tests on Juju
 
-First, make sure you have installed Juju and bootstrapped a Juju controller. You can provision a local controller on LXD and create a `k8s-e2e` model using:
+First, make sure you have installed Juju and bootstrapped a Juju controller. You can provision a local controller on LXD and create a `k8s-integration` model using:
 
 ```bash
 sudo snap install juju
 mkdir -p ~/.local/share
 juju bootstrap localhost
-juju add-model k8s-e2e
+juju add-model k8s-integration
 ```
 
 Then, run the tests with:
@@ -108,13 +108,13 @@ Then, run the tests with:
 ```bash
 export TEST_SNAP=$PWD/k8s.snap
 export TEST_SUBSTRATE=juju
-export TEST_JUJU_MODEL=k8s-e2e
+export TEST_JUJU_MODEL=k8s-integration
 
 export TEST_JUJU_CONTROLLER=localhost       # (optionally) specify Juju controller to use for running the tests
 export TEST_JUJU_BASE=ubuntu@22.04          # (optionally) specify base OS to use for new Juju machines
 export TEST_JUJU_CONSTRAINTS='mem=4G'       # (optionally) specify constraints for new Juju machines
 
-cd tests/e2e && tox -e e2e
+cd tests/integration && tox -e integration
 ```
 
 Alternatively, you can specify a list of existing Juju machines to use for the tests (e.g. machines created using `juju add-machine`):
@@ -122,11 +122,11 @@ Alternatively, you can specify a list of existing Juju machines to use for the t
 ```bash
 export TEST_SNAP=$PWD/k8s.snap
 export TEST_SUBSTRATE=juju
-export TEST_JUJU_MODEL=k8s-e2e
+export TEST_JUJU_MODEL=k8s-integration
 
 export TEST_JUJU_MACHINES=0,1,2
 
-cd tests/e2e && tox -e e2e
+cd tests/integration && tox -e integration
 ```
 
 ## Writing an End to End test
@@ -138,13 +138,13 @@ Make sure to use the [Harness](./tests/conftest.py) fixture. That way, there _sh
 A typical end to end test for feature `<feature>` should look like this:
 
 ```python
-# tests/e2e/tests/test_<feature>.py
+# tests/integration/tests/test_<feature>.py
 #
 # Copyright 2024 Canonical, Ltd.
 #
 import logging
 
-from e2e_util import harness, util
+from test_util import harness, util
 
 LOG = logging.getLogger(__name__)
 FEATURE_NODE_COUNT = 3  # number of machines necessary for the test
