+++ tests/integration/templates/loadbalancer-test.yaml
