+++ tests/integration/templates/nginx-pod.yaml
