+++ tests/integration/templates/storage-setup.yaml
