+++ tests/integration/templates/storage-test.yaml
