+++ tests/integration/tests/test_cilium_e2e.py
@@ -6,7 +6,7 @@ import platform
 from typing import List
 
 import pytest
-from e2e_util import config, harness, util
+from test_util import config, harness, util
 
 LOG = logging.getLogger(__name__)
