+++ tests/integration/tests/test_metrics_server.py
@@ -5,7 +5,7 @@ import logging
 from typing import List
 
 import pytest
-from e2e_util import config, harness, util
+from test_util import config, harness, util
 
 LOG = logging.getLogger(__name__)
