+++ tests/integration/tests/test_network.py
@@ -6,8 +6,8 @@ import logging
 from pathlib import Path
 from typing import List
 
-from e2e_util import harness, util
-from e2e_util.config import MANIFESTS_DIR
+from test_util import harness, util
+from test_util.config import MANIFESTS_DIR
 
 LOG = logging.getLogger(__name__)
