+++ tests/integration/tests/test_storage.py
@@ -8,8 +8,8 @@ from pathlib import Path
 from typing import List
 
 import pytest
-from e2e_util import config, harness, util
-from e2e_util.config import MANIFESTS_DIR
+from test_util import config, harness, util
+from test_util.config import MANIFESTS_DIR
 
 LOG = logging.getLogger(__name__)
