+++ tests/integration/tests/test_util/harness/__init__.py
@@ -1,0 +1,18 @@
+#
+# Copyright 2024 Canonical, Ltd.
+#
+from test_util.harness.base import Harness, HarnessError, Instance
+from test_util.harness.juju import JujuHarness
+from test_util.harness.local import LocalHarness
+from test_util.harness.lxd import LXDHarness
+from test_util.harness.multipass import MultipassHarness
+
+__all__ = [
+    HarnessError,
+    Harness,
+    Instance,
+    JujuHarness,
+    LocalHarness,
+    LXDHarness,
+    MultipassHarness,
+]
