+++ tests/integration/tests/test_util/harness/juju.py
@@ -7,9 +7,9 @@ import shlex
 import subprocess
 from pathlib import Path
 
-from e2e_util import config
-from e2e_util.harness import Harness, HarnessError, Instance
-from e2e_util.util import run
+from test_util import config
+from test_util.harness import Harness, HarnessError, Instance
+from test_util.util import run
 
 LOG = logging.getLogger(__name__)
