+++ tests/integration/tests/test_util/harness/local.py
@@ -9,8 +9,8 @@ import socket
 import subprocess
 from pathlib import Path
 
-from e2e_util.harness import Harness, HarnessError, Instance
-from e2e_util.util import run
+from test_util.harness import Harness, HarnessError, Instance
+from test_util.util import run
 
 LOG = logging.getLogger(__name__)
