+++ tests/integration/tox.ini
@@ -1,7 +1,7 @@
 [tox]
 no_package = True
 skip_missing_interpreters = True
-env_list = format, lint, e2e
+env_list = format, lint, integration
 min_version = 4.0.0
 
 [testenv]
@@ -29,7 +29,7 @@ commands =
     isort {tox_root}/tests --profile=black --check
     black {tox_root}/tests --check --diff
 
-[testenv:e2e]
+[testenv:integration]
 description = Run integration tests
 deps =
     -r {tox_root}/requirements-test.txt
